Relocate FilePathUtils to src/utils module
