+++ src/extensions/default/JavaScriptCodeHints/ScopeManager.js
@@ -44,7 +44,7 @@ define(function (require, exports, module) {
         ExtensionUtils      = brackets.getModule("utils/ExtensionUtils"),
         FileSystem          = brackets.getModule("filesystem/FileSystem"),
         FileUtils           = brackets.getModule("file/FileUtils"),
-        FilePathUtils       = brackets.getModule("file/FilePathUtils"),
+        FilePathUtils       = brackets.getModule("utils/FilePathUtils"),
         LanguageManager     = brackets.getModule("language/LanguageManager"),
         PreferencesManager  = brackets.getModule("preferences/PreferencesManager"),
         ProjectManager      = brackets.getModule("project/ProjectManager"),
