+++ src/extensions/default/QuickView/main.js
@@ -33,7 +33,7 @@ define(function (require, exports, module) {
         CSSUtils            = brackets.getModule("language/CSSUtils"),
         EditorManager       = brackets.getModule("editor/EditorManager"),
         ExtensionUtils      = brackets.getModule("utils/ExtensionUtils"),
-        FilePathUtils       = brackets.getModule("file/FilePathUtils"),
+        FilePathUtils       = brackets.getModule("utils/FilePathUtils"),
         Menus               = brackets.getModule("command/Menus"),
         PreferencesManager  = brackets.getModule("preferences/PreferencesManager"),
         Strings             = brackets.getModule("strings"),
