+++ src/search/SearchModel.js
@@ -26,7 +26,7 @@
 define(function (require, exports, module) {
     "use strict";
     
-    var FilePathUtils   = require("file/FilePathUtils"),
+    var FilePathUtils   = require("utils/FilePathUtils"),
         EventDispatcher = require("utils/EventDispatcher"),
         FindUtils   = require("search/FindUtils");
